+++ plugin-dev/Scripts/upload-debug-symbols.sh
@@ -72,12 +72,6 @@ if [ ! -z $ENABLED_CONFIGS ]; then
     fi
 fi
 
-export SENTRY_PROPERTIES="$projectPath/sentry.properties"
-if [ ! -f "$SENTRY_PROPERTIES" ]; then
-    echo "Sentry: Properties file is missing: '$SENTRY_PROPERTIES'"
-    exit
-fi
-
 if [ $targetPlatform = "IOS" ] || [ $targetPlatform = "Mac" ]; then
     SENTRY_CLI_EXEC="$pluginPath/Source/ThirdParty/CLI/sentry-cli-Darwin-universal"
 elif [ $targetPlatform = "Linux" ] || [ $targetPlatform = "LinuxArm64" ]; then
@@ -92,7 +86,28 @@ if [ ! -f "$SENTRY_CLI_EXEC" ]; then
     exit
 fi
 
-echo "Sentry: Upload started using PropertiesFile '$SENTRY_PROPERTIES'"
+PROPERTIES_FILE="$projectPath/sentry.properties"
+
+echo "Sentry: Looking for properties file '$SENTRY_PROPERTIES'"
+
+if [ -f "$PROPERTIES_FILE" ]; then
+    echo "Sentry: Properties file found. Starting upload..."
+    export SENTRY_PROPERTIES=$PROPERTIES_FILE
+else
+    echo "Sentry: Properties file not found. Falling back to environment variables."
+    if [ -z "$SENTRY_PROJECT" ]; then        
+        echo "Error: SENTRY_PROJECT env var is not set. Skipping..."
+        exit
+    fi
+    if [ -z "$SENTRY_ORG" ]; then        
+        echo "Error: SENTRY_ORG env var is not set. Skipping..."
+        exit
+    fi
+    if [ -z "$SENTRY_AUTH_TOKEN" ]; then        
+        echo "Error: SENTRY_AUTH_TOKEN env var is not set. Skipping..."
+        exit
+    fi
+fi
 
 chmod +x "$SENTRY_CLI_EXEC"