+++ plugin-dev/Source/Sentry/Private/SentrySubsystem.cpp
@@ -7,7 +7,6 @@
 #include "SentryBreadcrumb.h"
 #include "SentryDefines.h"
 #include "SentryEvent.h"
-#include "SentryId.h"
 #include "SentryUser.h"
 #include "SentryUserFeedback.h"
 #include "SentryBeforeSendHandler.h"
@@ -16,7 +15,7 @@
 #include "SentryTransaction.h"
 #include "SentryTransactionContext.h"
 #include "SentryOutputDevice.h"
-#include "SentryOutputDeviceError.h"
+#include "SentryErrorOutputDevice.h"
 
 #include "CoreGlobals.h"
 #include "Engine/World.h"
@@ -39,8 +38,9 @@ void USentrySubsystem::Initialize(FSubsystemCollectionBase& Collection)
 	SubsystemNativeImpl = MakeShareable(new FPlatformSentrySubsystem());
 
 	const USentrySettings* Settings = FSentryModule::Get().GetSettings();
+	check(Settings);
 
-	UE_LOG(LogSentrySdk, Log, TEXT("Sentry plugin auto initialization: %s"), Settings->InitAutomatically ? TEXT("true") : TEXT("false"));
+	UE_LOG(LogSentrySdk, Log, TEXT("Sentry plugin auto initialization: %s"), *LexToString(Settings->InitAutomatically));
 
 	if (Settings->InitAutomatically)
 	{
@@ -59,9 +59,11 @@ void USentrySubsystem::Deinitialize()
 
 void USentrySubsystem::Initialize()
 {
+	check(SubsystemNativeImpl);
+
 	if (!SubsystemNativeImpl)
 	{
-		UE_LOG(LogSentrySdk, Warning, TEXT("Sentry subsystem is invalid and can't be initialized."));
+		UE_LOG(LogSentrySdk, Error, TEXT("Sentry subsystem is invalid and can't be initialized."));
 		return;
 	}
 
@@ -72,20 +74,21 @@ void USentrySubsystem::Initialize()
 	}
 
 	const USentrySettings* Settings = FSentryModule::Get().GetSettings();
+	check(Settings);
 
-	if(Settings->Dsn.IsEmpty())
+	if (Settings->Dsn.IsEmpty())
 	{
 		UE_LOG(LogSentrySdk, Warning, TEXT("Sentry requires minimal configuration for its initialization - please provide the DSN in plugin settings."));
 		return;
 	}
 
-	if(!IsCurrentBuildConfigurationEnabled() || !IsCurrentBuildTargetEnabled())
+	if (!IsCurrentBuildConfigurationEnabled() || !IsCurrentBuildTargetEnabled())
 	{
 		UE_LOG(LogSentrySdk, Warning, TEXT("Sentry initialization skipped since event capturing is disabled for the current configuration/target/build in plugin settings."));
 		return;
 	}
 
-	if(IsPromotedBuildsOnlyEnabled() && !FApp::GetEngineIsPromotedBuild())
+	if (IsPromotedBuildsOnlyEnabled() && !FApp::GetEngineIsPromotedBuild())
 	{
 		UE_LOG(LogSentrySdk, Warning, TEXT("Sentry initialization skipped since event capturing is disabled for the non-promoted builds in plugin settings."));
 		return;
@@ -96,6 +99,7 @@ void USentrySubsystem::Initialize()
 		: USentryBeforeSendHandler::StaticClass();
 
 	BeforeSendHandler = NewObject<USentryBeforeSendHandler>(this, BeforeSendHandlerClass);
+	check(BeforeSendHandler);
 
 	BeforeBreadcrumbHandler = Settings->BeforeBreadcrumbHandler != nullptr
 		? NewObject<USentryBeforeBreadcrumbHandler>(this, static_cast<UClass*>(Settings->BeforeBreadcrumbHandler))
@@ -106,10 +110,11 @@ void USentrySubsystem::Initialize()
 		: USentryTraceSampler::StaticClass();
 
 	TraceSampler = NewObject<USentryTraceSampler>(this, TraceSamplerClass);
+	check(TraceSampler);
 
 	SubsystemNativeImpl->InitWithSettings(Settings, BeforeSendHandler, BeforeBreadcrumbHandler, TraceSampler);
 
-	if(!SubsystemNativeImpl->IsEnabled())
+	if (!SubsystemNativeImpl->IsEnabled())
 	{
 		UE_LOG(LogSentrySdk, Error, TEXT("Sentry initialization failed."));
 		return;
@@ -126,21 +131,12 @@ void USentrySubsystem::Initialize()
 	ConfigureBreadcrumbs();
 
 	ConfigureOutputDevice();
+	ConfigureErrorOutputDevice();
 
-#if PLATFORM_WINDOWS
-	if (FEngineVersion::Current().GetMajor() == 5 && FEngineVersion::Current().GetMinor() >= 2)
+	OnEnsureDelegate = FCoreDelegates::OnHandleSystemEnsure.AddWeakLambda(this, [this]()
 	{
-		if (Settings->EnableAutoCrashCapturing)
-		{
-			ConfigureOutputDeviceError();
-		}
-	}
-#else
-	ConfigureOutputDeviceError();
-#endif
+		verify(SubsystemNativeImpl);
 
-	OnEnsureDelegate = FCoreDelegates::OnHandleSystemEnsure.AddLambda([this]()
-	{
 		FString EnsureMessage = GErrorHist;
 		SubsystemNativeImpl->CaptureEnsure(TEXT("Ensure failed"), EnsureMessage.TrimStartAndEnd());
 	});
@@ -149,6 +145,7 @@ void USentrySubsystem::Initialize()
 void USentrySubsystem::InitializeWithSettings(const FConfigureSettingsDelegate& OnConfigureSettings)
 {
 	USentrySettings* Settings = FSentryModule::Get().GetSettings();
+	check(Settings);
 
 	OnConfigureSettings.ExecuteIfBound(Settings);
 
@@ -157,15 +154,15 @@ void USentrySubsystem::InitializeWithSettings(const FConfigureSettingsDelegate& 
 
 void USentrySubsystem::Close()
 {
-	if(GLog && OutputDevice)
+	if (GLog && OutputDevice)
 	{
 		GLog->RemoveOutputDevice(OutputDevice.Get());
 		OutputDevice = nullptr;
 	}
 
-	if(GError && OutputDeviceError)
+	if (GError && OutputDeviceError)
 	{
-		if(OnAssertDelegate.IsValid())
+		if (OnAssertDelegate.IsValid())
 		{
 			OutputDeviceError->OnAssert.Remove(OnAssertDelegate);
 			OnAssertDelegate.Reset();
@@ -175,127 +172,169 @@ void USentrySubsystem::Close()
 		OutputDeviceError = nullptr;
 	}
 
-	if(OnEnsureDelegate.IsValid())
+	if (OnEnsureDelegate.IsValid())
 	{
 		FCoreDelegates::OnHandleSystemEnsure.Remove(OnEnsureDelegate);
 		OnEnsureDelegate.Reset();
 	}
 
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return;
+	}
 
 	SubsystemNativeImpl->Close();
 }
 
-bool USentrySubsystem::IsEnabled()
+bool USentrySubsystem::IsEnabled() const
 {
-	if (!SubsystemNativeImpl)
-		return false;
-
-	return SubsystemNativeImpl->IsEnabled();
+	return SubsystemNativeImpl ? SubsystemNativeImpl->IsEnabled() : false;
 }
 
-ESentryCrashedLastRun USentrySubsystem::IsCrashedLastRun()
+ESentryCrashedLastRun USentrySubsystem::IsCrashedLastRun() const
 {
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return ESentryCrashedLastRun::NotEvaluated;
+	}
 
 	return SubsystemNativeImpl->IsCrashedLastRun();
 }
 
 void USentrySubsystem::AddBreadcrumb(USentryBreadcrumb* Breadcrumb)
 {
+	check(SubsystemNativeImpl);
+	check(Breadcrumb);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return;
+	}
 
 	SubsystemNativeImpl->AddBreadcrumb(Breadcrumb->GetNativeObject());
 }
 
 void USentrySubsystem::AddBreadcrumbWithParams(const FString& Message, const FString& Category, const FString& Type, const TMap<FString, FString>& Data, ESentryLevel Level)
 {
+	check(SubsystemNativeImpl);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return;
+	}
 
 	SubsystemNativeImpl->AddBreadcrumbWithParams(Message, Category, Type, Data, Level);
 }
 
 void USentrySubsystem::ClearBreadcrumbs()
 {
+	check(SubsystemNativeImpl);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return;
+	}
 
 	SubsystemNativeImpl->ClearBreadcrumbs();
 }
 
-USentryId* USentrySubsystem::CaptureMessage(const FString& Message, ESentryLevel Level)
+FString USentrySubsystem::CaptureMessage(const FString& Message, ESentryLevel Level)
 {
+	check(SubsystemNativeImpl);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
-		return nullptr;
+	{
+		return FString();
+	}
 
-	TSharedPtr<ISentryId> idNativeImpl = SubsystemNativeImpl->CaptureMessage(Message, Level);
+	TSharedPtr<ISentryId> SentryId = SubsystemNativeImpl->CaptureMessage(Message, Level);
 
-	return USentryId::Create(idNativeImpl);
+	return SentryId->ToString();
 }
 
-USentryId* USentrySubsystem::CaptureMessageWithScope(const FString& Message, const FConfigureScopeDelegate& OnConfigureScope, ESentryLevel Level)
+FString USentrySubsystem::CaptureMessageWithScope(const FString& Message, const FConfigureScopeDelegate& OnConfigureScope, ESentryLevel Level)
 {
 	return CaptureMessageWithScope(Message, FConfigureScopeNativeDelegate::CreateUFunction(const_cast<UObject*>(OnConfigureScope.GetUObject()), OnConfigureScope.GetFunctionName()), Level);
 }
 
-USentryId* USentrySubsystem::CaptureMessageWithScope(const FString& Message, const FConfigureScopeNativeDelegate& OnConfigureScope, ESentryLevel Level)
+FString USentrySubsystem::CaptureMessageWithScope(const FString& Message, const FConfigureScopeNativeDelegate& OnConfigureScope, ESentryLevel Level)
 {
-	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
-		return nullptr;
+	check(SubsystemNativeImpl);
 
-	TSharedPtr<ISentryId> idNativeImpl = SubsystemNativeImpl->CaptureMessageWithScope(Message, FSentryScopeDelegate::CreateLambda([&](TSharedPtr<ISentryScope> nativeScope)
+	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
 	{
-		USentryScope* unrealScope = USentryScope::Create(nativeScope);
-		OnConfigureScope.ExecuteIfBound(unrealScope);
+		return FString();
+	}
+
+	TSharedPtr<ISentryId> SentryId = SubsystemNativeImpl->CaptureMessageWithScope(Message, FSentryScopeDelegate::CreateLambda([OnConfigureScope](TSharedPtr<ISentryScope> NativeScope)
+	{
+		USentryScope* UnrealScope = USentryScope::Create(NativeScope);
+		OnConfigureScope.ExecuteIfBound(UnrealScope);
 	}), Level);
 
-	return USentryId::Create(idNativeImpl);
+	return SentryId->ToString();
 }
 
-USentryId* USentrySubsystem::CaptureEvent(USentryEvent* Event)
+FString USentrySubsystem::CaptureEvent(USentryEvent* Event)
 {
+	check(SubsystemNativeImpl);
+	check(Event);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
-		return nullptr;
+	{
+		return FString();
+	}
 
-	TSharedPtr<ISentryId> idNativeImpl = SubsystemNativeImpl->CaptureEvent(Event->GetNativeObject());
+	TSharedPtr<ISentryId> SentryId = SubsystemNativeImpl->CaptureEvent(Event->GetNativeObject());
 
-	return USentryId::Create(idNativeImpl);
+	return SentryId->ToString();
 }
 
-USentryId* USentrySubsystem::CaptureEventWithScope(USentryEvent* Event, const FConfigureScopeDelegate& OnConfigureScope)
+FString USentrySubsystem::CaptureEventWithScope(USentryEvent* Event, const FConfigureScopeDelegate& OnConfigureScope)
 {
 	return CaptureEventWithScope(Event, FConfigureScopeNativeDelegate::CreateUFunction(const_cast<UObject*>(OnConfigureScope.GetUObject()), OnConfigureScope.GetFunctionName()));
 }
 
-USentryId* USentrySubsystem::CaptureEventWithScope(USentryEvent* Event, const FConfigureScopeNativeDelegate& OnConfigureScope)
+FString USentrySubsystem::CaptureEventWithScope(USentryEvent* Event, const FConfigureScopeNativeDelegate& OnConfigureScope)
 {
-	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
-		return nullptr;
+	check(SubsystemNativeImpl);
+	check(Event);
 
-	TSharedPtr<ISentryId> idNativeImpl = SubsystemNativeImpl->CaptureEventWithScope(Event->GetNativeObject(), FSentryScopeDelegate::CreateLambda([&](TSharedPtr<ISentryScope> nativeScope)
+	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
 	{
-		USentryScope* unrealScope = USentryScope::Create(nativeScope);
-		OnConfigureScope.ExecuteIfBound(unrealScope);
+		return FString();
+	}
+
+	TSharedPtr<ISentryId> SentryId = SubsystemNativeImpl->CaptureEventWithScope(Event->GetNativeObject(), FSentryScopeDelegate::CreateLambda([OnConfigureScope](TSharedPtr<ISentryScope> NativeScope)
+	{
+		USentryScope* UnrealScope = USentryScope::Create(NativeScope);
+		OnConfigureScope.ExecuteIfBound(UnrealScope);
 	}));
 
-	return USentryId::Create(idNativeImpl);
+	return SentryId->ToString();
 }
 
 void USentrySubsystem::CaptureUserFeedback(USentryUserFeedback* UserFeedback)
 {
+	check(SubsystemNativeImpl);
+	check(UserFeedback);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return;
+	}
 
 	SubsystemNativeImpl->CaptureUserFeedback(UserFeedback->GetNativeObject());
 }
 
-void USentrySubsystem::CaptureUserFeedbackWithParams(USentryId* EventId, const FString& Email, const FString& Comments, const FString& Name)
+void USentrySubsystem::CaptureUserFeedbackWithParams(const FString& EventId, const FString& Email, const FString& Comments, const FString& Name)
 {
-	USentryUserFeedback* UserFeedback = USentryUserFeedback::Create(CreateSharedSentryUserFeedback(EventId->GetNativeObject()));
+	check(SubsystemNativeImpl);
+	check(!EventId.IsEmpty());
+
+	USentryUserFeedback* UserFeedback = USentryUserFeedback::Create(CreateSharedSentryUserFeedback(EventId));
+	check(UserFeedback);
+
 	UserFeedback->SetEmail(Email);
 	UserFeedback->SetComment(Comments);
 	UserFeedback->SetName(Name);
@@ -305,16 +344,25 @@ void USentrySubsystem::CaptureUserFeedbackWithParams(USentryId* EventId, const F
 
 void USentrySubsystem::SetUser(USentryUser* User)
 {
+	check(SubsystemNativeImpl);
+	check(User);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return;
+	}
 
 	SubsystemNativeImpl->SetUser(User->GetNativeObject());
 }
 
 void USentrySubsystem::RemoveUser()
 {
+	check(SubsystemNativeImpl);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return;
+	}
 
 	SubsystemNativeImpl->RemoveUser();
 }
@@ -326,122 +374,178 @@ void USentrySubsystem::ConfigureScope(const FConfigureScopeDelegate& OnConfigure
 
 void USentrySubsystem::ConfigureScope(const FConfigureScopeNativeDelegate& OnConfigureScope)
 {
-	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
-		return;
+	check(SubsystemNativeImpl);
 
-	SubsystemNativeImpl->ConfigureScope(FSentryScopeDelegate::CreateLambda([&](TSharedPtr<ISentryScope> nativeScope)
+	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
 	{
-		USentryScope* unrealScope = USentryScope::Create(nativeScope);
-		OnConfigureScope.ExecuteIfBound(unrealScope);
+		return;
+	}
+
+	SubsystemNativeImpl->ConfigureScope(FSentryScopeDelegate::CreateLambda([OnConfigureScope](TSharedPtr<ISentryScope> NativeScope)
+	{
+		USentryScope* UnrealScope = USentryScope::Create(NativeScope);
+		OnConfigureScope.ExecuteIfBound(UnrealScope);
 	}));
 }
 
 void USentrySubsystem::SetContext(const FString& Key, const TMap<FString, FString>& Values)
 {
+	check(SubsystemNativeImpl);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return;
+	}
 
 	SubsystemNativeImpl->SetContext(Key, Values);
 }
 
 void USentrySubsystem::SetTag(const FString& Key, const FString& Value)
 {
+	check(SubsystemNativeImpl);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return;
+	}
 
 	SubsystemNativeImpl->SetTag(Key, Value);
 }
 
 void USentrySubsystem::RemoveTag(const FString& Key)
 {
+	check(SubsystemNativeImpl);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return;
+	}
 
 	SubsystemNativeImpl->RemoveTag(Key);
 }
 
 void USentrySubsystem::SetLevel(ESentryLevel Level)
 {
+	check(SubsystemNativeImpl);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return;
+	}
 
 	SubsystemNativeImpl->SetLevel(Level);
 }
 
 void USentrySubsystem::StartSession()
 {
+	check(SubsystemNativeImpl);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return;
+	}
 
 	SubsystemNativeImpl->StartSession();
 }
 
 void USentrySubsystem::EndSession()
 {
+	check(SubsystemNativeImpl);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return;
+	}
 
 	SubsystemNativeImpl->EndSession();
 }
 
 USentryTransaction* USentrySubsystem::StartTransaction(const FString& Name, const FString& Operation)
 {
+	check(SubsystemNativeImpl);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return nullptr;
+	}
 
-	TSharedPtr<ISentryTransaction> transactionNativeImpl = SubsystemNativeImpl->StartTransaction(Name, Operation);
+	TSharedPtr<ISentryTransaction> SentryTransaction = SubsystemNativeImpl->StartTransaction(Name, Operation);
+	check(SentryTransaction);
 
-	return USentryTransaction::Create(transactionNativeImpl);
+	return USentryTransaction::Create(SentryTransaction);
 }
 
 USentryTransaction* USentrySubsystem::StartTransactionWithContext(USentryTransactionContext* Context)
 {
+	check(SubsystemNativeImpl);
+	check(Context);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return nullptr;
+	}
 
-	TSharedPtr<ISentryTransaction> transactionNativeImpl = SubsystemNativeImpl->StartTransactionWithContext(Context->GetNativeObject());
+	TSharedPtr<ISentryTransaction> SentryTransaction = SubsystemNativeImpl->StartTransactionWithContext(Context->GetNativeObject());
+	check(SentryTransaction);
 
-	return USentryTransaction::Create(transactionNativeImpl);
+	return USentryTransaction::Create(SentryTransaction);
 }
 
 USentryTransaction* USentrySubsystem::StartTransactionWithContextAndTimestamp(USentryTransactionContext* Context, int64 Timestamp)
 {
+	check(SubsystemNativeImpl);
+	check(Context);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return nullptr;
+	}
 
-	TSharedPtr<ISentryTransaction> transactionNativeImpl = SubsystemNativeImpl->StartTransactionWithContextAndTimestamp(Context->GetNativeObject(), Timestamp);
+	TSharedPtr<ISentryTransaction> SentryTransaction = SubsystemNativeImpl->StartTransactionWithContextAndTimestamp(Context->GetNativeObject(), Timestamp);
+	check(SentryTransaction);
 
-	return USentryTransaction::Create(transactionNativeImpl);
+	return USentryTransaction::Create(SentryTransaction);
 }
 
 USentryTransaction* USentrySubsystem::StartTransactionWithContextAndOptions(USentryTransactionContext* Context, const TMap<FString, FString>& Options)
 {
+	check(SubsystemNativeImpl);
+	check(Context);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return nullptr;
+	}
 
-	TSharedPtr<ISentryTransaction> transactionNativeImpl = SubsystemNativeImpl->StartTransactionWithContextAndOptions(Context->GetNativeObject(), Options);
+	TSharedPtr<ISentryTransaction> SentryTransaction = SubsystemNativeImpl->StartTransactionWithContextAndOptions(Context->GetNativeObject(), Options);
+	check(SentryTransaction);
 
-	return USentryTransaction::Create(transactionNativeImpl);
+	return USentryTransaction::Create(SentryTransaction);
 }
 
 USentryTransactionContext* USentrySubsystem::ContinueTrace(const FString& SentryTrace, const TArray<FString>& BaggageHeaders)
 {
+	check(SubsystemNativeImpl);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return nullptr;
+	}
 
-	TSharedPtr<ISentryTransactionContext> transactionContextNativeImpl = SubsystemNativeImpl->ContinueTrace(SentryTrace, BaggageHeaders);
+	TSharedPtr<ISentryTransactionContext> SentryTransactionContext = SubsystemNativeImpl->ContinueTrace(SentryTrace, BaggageHeaders);
+	check(SentryTransactionContext);
 
-	return USentryTransactionContext::Create(transactionContextNativeImpl);
+	return USentryTransactionContext::Create(SentryTransactionContext);
 }
 
-bool USentrySubsystem::IsSupportedForCurrentSettings()
+bool USentrySubsystem::IsSupportedForCurrentSettings() const
 {
-	if(!IsCurrentBuildConfigurationEnabled() || !IsCurrentBuildTargetEnabled())
+	if (!IsCurrentBuildConfigurationEnabled() || !IsCurrentBuildTargetEnabled())
 	{
 		return false;
 	}
 
-	if(IsPromotedBuildsOnlyEnabled() && !FApp::GetEngineIsPromotedBuild())
+	if (IsPromotedBuildsOnlyEnabled() && !FApp::GetEngineIsPromotedBuild())
 	{
 		return false;
 	}
@@ -451,19 +555,23 @@ bool USentrySubsystem::IsSupportedForCurrentSettings()
 
 void USentrySubsystem::AddDefaultContext()
 {
+	check(SubsystemNativeImpl);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return;
+	}
 
 	TMap<FString, FString> DefaultContext;
 	DefaultContext.Add(TEXT("Engine version"), FEngineVersion::Current().ToString(EVersionComponent::Changelist));
 	DefaultContext.Add(TEXT("Plugin version"), FSentryModule::Get().GetPluginVersion());
-	DefaultContext.Add(TEXT("Is Marketplace version"), FSentryModule::Get().IsMarketplaceVersion() ? TEXT("True") : TEXT("False"));
+	DefaultContext.Add(TEXT("Is Marketplace version"), LexToString(FSentryModule::Get().IsMarketplaceVersion()));
 	DefaultContext.Add(TEXT("Configuration"), LexToString(FApp::GetBuildConfiguration()));
 	DefaultContext.Add(TEXT("Target Type"), LexToString(FApp::GetBuildTargetType()));
 	DefaultContext.Add(TEXT("Engine mode"), FGenericPlatformMisc::GetEngineMode());
-	DefaultContext.Add(TEXT("Is game"), FApp::IsGame() ? TEXT("True") : TEXT("False"));
-	DefaultContext.Add(TEXT("Is standalone"), FApp::IsStandalone() ? TEXT("True") : TEXT("False"));
-	DefaultContext.Add(TEXT("Is unattended"), FApp::IsUnattended() ? TEXT("True") : TEXT("False"));
+	DefaultContext.Add(TEXT("Is game"), LexToString(FApp::IsGame()));
+	DefaultContext.Add(TEXT("Is standalone"), LexToString(FApp::IsStandalone()));
+	DefaultContext.Add(TEXT("Is unattended"), LexToString(FApp::IsUnattended()));
 	DefaultContext.Add(TEXT("Game name"), FApp::GetName());
 
 	SubsystemNativeImpl->SetContext(TEXT("Unreal Engine"), DefaultContext);
@@ -471,8 +579,12 @@ void USentrySubsystem::AddDefaultContext()
 
 void USentrySubsystem::AddGpuContext()
 {
+	check(SubsystemNativeImpl);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return;
+	}
 
 	FGPUDriverInfo GpuDriverInfo = FPlatformMisc::GetGPUDriverInfo(FPlatformMisc::GetPrimaryGPUBrand());
 
@@ -486,8 +598,12 @@ void USentrySubsystem::AddGpuContext()
 
 void USentrySubsystem::AddDeviceContext()
 {
+	check(SubsystemNativeImpl);
+
 	if (!SubsystemNativeImpl || !SubsystemNativeImpl->IsEnabled())
+	{
 		return;
+	}
 
 	const FPlatformMemoryConstants& MemoryConstants = FPlatformMemory::GetConstants();
 
@@ -502,171 +618,171 @@ void USentrySubsystem::AddDeviceContext()
 
 void USentrySubsystem::PromoteTags()
 {
-	const USentrySettings* Settings = FSentryModule::Get().GetSettings();
+	check(SubsystemNativeImpl);
 
-	if(Settings->TagsPromotion.bPromoteBuildConfiguration)
+	const USentrySettings* Settings = FSentryModule::Get().GetSettings();
+	check(Settings);
+
+	if (Settings->TagsPromotion.bPromoteBuildConfiguration)
+	{
 		SubsystemNativeImpl->SetTag(TEXT("Configuration"), LexToString(FApp::GetBuildConfiguration()));
-	if(Settings->TagsPromotion.bPromoteTargetType)
+	}
+
+	if (Settings->TagsPromotion.bPromoteTargetType)
+	{
 		SubsystemNativeImpl->SetTag(TEXT("Target Type"), LexToString(FApp::GetBuildTargetType()));
-	if(Settings->TagsPromotion.bPromoteEngineMode)
+	}
+
+	if (Settings->TagsPromotion.bPromoteEngineMode)
+	{
 		SubsystemNativeImpl->SetTag(TEXT("Engine Mode"), FGenericPlatformMisc::GetEngineMode());
-	if(Settings->TagsPromotion.bPromoteIsGame)
-		SubsystemNativeImpl->SetTag(TEXT("Is game"), FApp::IsGame() ? TEXT("True") : TEXT("False"));
-	if(Settings->TagsPromotion.bPromoteIsStandalone)
-		SubsystemNativeImpl->SetTag(TEXT("Is standalone"), FApp::IsStandalone() ? TEXT("True") : TEXT("False"));
-	if(Settings->TagsPromotion.bPromoteIsUnattended)
-		SubsystemNativeImpl->SetTag(TEXT("Is unattended"), FApp::IsUnattended() ? TEXT("True") : TEXT("False"));
+	}
+
+	if (Settings->TagsPromotion.bPromoteIsGame)
+	{
+		SubsystemNativeImpl->SetTag(TEXT("Is game"), LexToString(FApp::IsGame()));
+	}
+
+	if (Settings->TagsPromotion.bPromoteIsStandalone)
+	{
+		SubsystemNativeImpl->SetTag(TEXT("Is standalone"), LexToString(FApp::IsStandalone()));
+	}
+
+	if (Settings->TagsPromotion.bPromoteIsUnattended)
+	{
+		SubsystemNativeImpl->SetTag(TEXT("Is unattended"), LexToString(FApp::IsUnattended()));
+	}
 }
 
 void USentrySubsystem::ConfigureBreadcrumbs()
 {
 	const USentrySettings* Settings = FSentryModule::Get().GetSettings();
+	check(Settings);
 
-	if(Settings->AutomaticBreadcrumbs.bOnMapLoadingStarted)
+	if (Settings->AutomaticBreadcrumbs.bOnMapLoadingStarted)
 	{
-		PreLoadMapDelegate = FCoreUObjectDelegates::PreLoadMap.AddLambda([this](const FString& MapName)
+		PreLoadMapDelegate = FCoreUObjectDelegates::PreLoadMap.AddWeakLambda(this, [this](const FString& MapName)
 		{
-			AddBreadcrumbWithParams(TEXT("PreLoadMap"), TEXT("Unreal"), TEXT("Default"),
-				{{TEXT("Map"), MapName}}, ESentryLevel::Info);
+			AddBreadcrumbWithParams(TEXT("PreLoadMap"), TEXT("Unreal"), TEXT("Default"), {{TEXT("Map"), MapName}}, ESentryLevel::Info);
 		});
 	}
 
-	if(Settings->AutomaticBreadcrumbs.bOnMapLoaded)
+	if (Settings->AutomaticBreadcrumbs.bOnMapLoaded)
 	{
-		PostLoadMapDelegate = FCoreUObjectDelegates::PostLoadMapWithWorld.AddLambda([this](UWorld* World)
+		PostLoadMapDelegate = FCoreUObjectDelegates::PostLoadMapWithWorld.AddWeakLambda(this, [this](UWorld* World)
 		{
 			if (World)
 			{
-				AddBreadcrumbWithParams(TEXT("PostLoadMapWithWorld"), TEXT("Unreal"), TEXT("Default"),
-					{{TEXT("Map"), World->GetMapName()}}, ESentryLevel::Info);
+				AddBreadcrumbWithParams(TEXT("PostLoadMapWithWorld"), TEXT("Unreal"), TEXT("Default"), {{TEXT("Map"), World->GetMapName()}}, ESentryLevel::Info);
 			}
 			else
 			{
-				AddBreadcrumbWithParams(TEXT("PostLoadMapWithWorld"), TEXT("Unreal"), TEXT("Default"),
-					{{TEXT("Error"), TEXT("Map load failed")}}, ESentryLevel::Error);
+				AddBreadcrumbWithParams(TEXT("PostLoadMapWithWorld"), TEXT("Unreal"), TEXT("Default"), {{TEXT("Error"), TEXT("Map load failed")}}, ESentryLevel::Error);
 			}
 		});
 	}
 
-	if(Settings->AutomaticBreadcrumbs.bOnGameStateClassChanged)
+	if (Settings->AutomaticBreadcrumbs.bOnGameStateClassChanged)
 	{
-		GameStateChangedDelegate = FCoreDelegates::GameStateClassChanged.AddLambda([this](const FString& GameState)
+		GameStateChangedDelegate = FCoreDelegates::GameStateClassChanged.AddWeakLambda(this, [this](const FString& GameState)
 		{
-			AddBreadcrumbWithParams(TEXT("GameStateClassChanged"), TEXT("Unreal"), TEXT("Default"),
-				{{TEXT("GameState"), GameState}}, ESentryLevel::Info);
+			AddBreadcrumbWithParams(TEXT("GameStateClassChanged"), TEXT("Unreal"), TEXT("Default"), {{TEXT("GameState"), GameState}}, ESentryLevel::Info);
 		});
 	}
 
-	if(Settings->AutomaticBreadcrumbs.bOnUserActivityStringChanged)
+	if (Settings->AutomaticBreadcrumbs.bOnUserActivityStringChanged)
 	{
-		UserActivityChangedDelegate = FCoreDelegates::UserActivityStringChanged.AddLambda([this](const FString& Activity)
+		UserActivityChangedDelegate = FCoreDelegates::UserActivityStringChanged.AddWeakLambda(this, [this](const FString& Activity)
 		{
-			AddBreadcrumbWithParams(TEXT("UserActivityStringChanged"), TEXT("Unreal"), TEXT("Default"),
-				{{TEXT("Activity"), Activity}}, ESentryLevel::Info);
+			AddBreadcrumbWithParams(TEXT("UserActivityStringChanged"), TEXT("Unreal"), TEXT("Default"), {{TEXT("Activity"), Activity}}, ESentryLevel::Info);
 		});
 	}
 
-	if(Settings->AutomaticBreadcrumbs.bOnGameSessionIDChanged)
+	if (Settings->AutomaticBreadcrumbs.bOnGameSessionIDChanged)
 	{
-		GameSessionIDChangedDelegate = FCoreDelegates::GameSessionIDChanged.AddLambda([this](const FString& SessionId)
+		GameSessionIDChangedDelegate = FCoreDelegates::GameSessionIDChanged.AddWeakLambda(this, [this](const FString& SessionId)
 		{
-			AddBreadcrumbWithParams(TEXT("GameSessionIDChanged"), TEXT("Unreal"), TEXT("Default"),
-				{{TEXT("Session ID"), SessionId}}, ESentryLevel::Info);
+			AddBreadcrumbWithParams(TEXT("GameSessionIDChanged"), TEXT("Unreal"), TEXT("Default"), {{TEXT("Session ID"), SessionId}}, ESentryLevel::Info);
 		});
 	}
 }
 
 void USentrySubsystem::DisableAutomaticBreadcrumbs()
 {
-	if(PreLoadMapDelegate.IsValid())
+	if (PreLoadMapDelegate.IsValid())
 	{
 		FCoreUObjectDelegates::PreLoadMap.Remove(PreLoadMapDelegate);
 	}
 
-	if(PostLoadMapDelegate.IsValid())
+	if (PostLoadMapDelegate.IsValid())
 	{
 		FCoreUObjectDelegates::PostLoadMapWithWorld.Remove(PostLoadMapDelegate);
 	}
 
-	if(GameStateChangedDelegate.IsValid())
+	if (GameStateChangedDelegate.IsValid())
 	{
 		FCoreDelegates::GameStateClassChanged.Remove(GameStateChangedDelegate);
 	}
 
-	if(UserActivityChangedDelegate.IsValid())
+	if (UserActivityChangedDelegate.IsValid())
 	{
 		FCoreDelegates::UserActivityStringChanged.Remove(UserActivityChangedDelegate);
 	}
 
-	if(GameSessionIDChangedDelegate.IsValid())
+	if (GameSessionIDChangedDelegate.IsValid())
 	{
 		FCoreDelegates::GameSessionIDChanged.Remove(GameSessionIDChangedDelegate);
 	}
 }
 
-bool USentrySubsystem::IsCurrentBuildConfigurationEnabled()
+bool USentrySubsystem::IsCurrentBuildConfigurationEnabled() const
 {
 	const USentrySettings* Settings = FSentryModule::Get().GetSettings();
-
-	bool IsBuildConfigurationEnabled;
+	check(Settings);
 
 	switch (FApp::GetBuildConfiguration())
 	{
 	case EBuildConfiguration::Debug:
-		IsBuildConfigurationEnabled = Settings->EnableBuildConfigurations.bEnableDebug;
-		break;
+		return Settings->EnableBuildConfigurations.bEnableDebug;
 	case EBuildConfiguration::DebugGame:
-		IsBuildConfigurationEnabled = Settings->EnableBuildConfigurations.bEnableDebugGame;
-		break;
+		return Settings->EnableBuildConfigurations.bEnableDebugGame;
 	case EBuildConfiguration::Development:
-		IsBuildConfigurationEnabled = Settings->EnableBuildConfigurations.bEnableDevelopment;
-		break;
+		return Settings->EnableBuildConfigurations.bEnableDevelopment;
 	case EBuildConfiguration::Shipping:
-		IsBuildConfigurationEnabled = Settings->EnableBuildConfigurations.bEnableShipping;
-		break;
+		return Settings->EnableBuildConfigurations.bEnableShipping;
 	case EBuildConfiguration::Test:
-		IsBuildConfigurationEnabled = Settings->EnableBuildConfigurations.bEnableTest;
-		break;
+		return Settings->EnableBuildConfigurations.bEnableTest;
 	default:
-		IsBuildConfigurationEnabled = false;
+		return false;
 	}
-
-	return IsBuildConfigurationEnabled;
 }
 
-bool USentrySubsystem::IsCurrentBuildTargetEnabled()
+bool USentrySubsystem::IsCurrentBuildTargetEnabled() const
 {
 	const USentrySettings* Settings = FSentryModule::Get().GetSettings();
-
-	bool IsBuildTargetTypeEnabled;
+	check(Settings);
 
 	switch (FApp::GetBuildTargetType())
 	{
 	case EBuildTargetType::Game:
-		IsBuildTargetTypeEnabled = Settings->EnableBuildTargets.bEnableGame;
-		break;
+		return Settings->EnableBuildTargets.bEnableGame;
 	case EBuildTargetType::Server:
-		IsBuildTargetTypeEnabled = Settings->EnableBuildTargets.bEnableServer;
-		break;
+		return Settings->EnableBuildTargets.bEnableServer;
 	case EBuildTargetType::Client:
-		IsBuildTargetTypeEnabled = Settings->EnableBuildTargets.bEnableClient;
-		break;
+		return Settings->EnableBuildTargets.bEnableClient;
 	case EBuildTargetType::Editor:
-		IsBuildTargetTypeEnabled = Settings->EnableBuildTargets.bEnableEditor;
-		break;
+		return Settings->EnableBuildTargets.bEnableEditor;
 	case EBuildTargetType::Program:
-		IsBuildTargetTypeEnabled = Settings->EnableBuildTargets.bEnableProgram;
-		break;
+		return Settings->EnableBuildTargets.bEnableProgram;
 	default:
-		IsBuildTargetTypeEnabled = false;
+		return false;
 	}
-
-	return IsBuildTargetTypeEnabled;
 }
 
-bool USentrySubsystem::IsPromotedBuildsOnlyEnabled()
+bool USentrySubsystem::IsPromotedBuildsOnlyEnabled() const
 {
 	const USentrySettings* Settings = FSentryModule::Get().GetSettings();
+	check(Settings);
 
 	return Settings->EnableForPromotedBuildsOnly;
 }
@@ -681,18 +797,18 @@ void USentrySubsystem::ConfigureOutputDevice()
 	}
 }
 
-void USentrySubsystem::ConfigureOutputDeviceError()
+void USentrySubsystem::ConfigureErrorOutputDevice()
 {
-	OutputDeviceError = MakeShareable(new FSentryOutputDeviceError(GError));
+	OutputDeviceError = MakeShareable(new FSentryErrorOutputDevice(GError));
 	if (OutputDeviceError)
 	{
-		OnAssertDelegate = OutputDeviceError->OnAssert.AddLambda([this](const FString& Message)
+#if PLATFORM_ANDROID
+		OnAssertDelegate = OutputDeviceError->OnAssert.AddWeakLambda(this, [this](const FString& Message)
 		{
-		#if PLATFORM_ANDROID
 			GError->HandleError();
 			PLATFORM_BREAK();
-		#endif
 		});
+#endif // PLATFORM_ANDROID
 
 		GError = OutputDeviceError.Get();
 	}