+++ plugin-dev/Source/Sentry/Public/SentrySubsystem.h
@@ -13,7 +13,6 @@
 class USentrySettings;
 class USentryBreadcrumb;
 class USentryEvent;
-class USentryId;
 class USentryUserFeedback;
 class USentryUser;
 class USentryBeforeSendHandler;
@@ -24,7 +23,7 @@ class USentryTransactionContext;
 
 class ISentrySubsystem;
 class FSentryOutputDevice;
-class FSentryOutputDeviceError;
+class FSentryErrorOutputDevice;
 
 DECLARE_DYNAMIC_DELEGATE_OneParam(FConfigureSettingsDelegate, USentrySettings*, Settings);
 
@@ -56,11 +55,11 @@ public:
 
 	/** Checks whether the Sentry SDK was initialized and event capturing is enabled. */
 	UFUNCTION(BlueprintPure, Category = "Sentry")
-	bool IsEnabled();
+	bool IsEnabled() const;
 
 	/** Checks whether the app crashed during the last run. */
 	UFUNCTION(BlueprintPure, Category = "Sentry")
-	ESentryCrashedLastRun IsCrashedLastRun();
+	ESentryCrashedLastRun IsCrashedLastRun() const;
 
 	/**
 	 * Adds a breadcrumb to the current Scope.
@@ -107,9 +106,11 @@ public:
 	 *
 	 * @param Message The message to send.
 	 * @param Level The message level.
+	 *
+	 * @return Event ID (non-empty if successful)
 	 */
 	UFUNCTION(BlueprintCallable, Category = "Sentry")
-	USentryId* CaptureMessage(const FString& Message, ESentryLevel Level = ESentryLevel::Info);
+	FString CaptureMessage(const FString& Message, ESentryLevel Level = ESentryLevel::Info);
 
 	/**
 	 * Captures the message with a configurable scope.
@@ -120,19 +121,21 @@ public:
 	 * @param OnConfigureScope The callback to configure the scope.
 	 * @param Level The message level.
 	 *
-	 * @note: Not supported for Windows/Linux.
+	 * @return Event ID (non-empty if successful)
 	 */
 	UFUNCTION(BlueprintCallable, Category = "Sentry", meta = (AutoCreateRefTerm = "OnConfigureScope"))
-	USentryId* CaptureMessageWithScope(const FString& Message, const FConfigureScopeDelegate& OnConfigureScope, ESentryLevel Level = ESentryLevel::Info);
-	USentryId* CaptureMessageWithScope(const FString& Message, const FConfigureScopeNativeDelegate& OnConfigureScope, ESentryLevel Level = ESentryLevel::Info);
+	FString CaptureMessageWithScope(const FString& Message, const FConfigureScopeDelegate& OnConfigureScope, ESentryLevel Level = ESentryLevel::Info);
+	FString CaptureMessageWithScope(const FString& Message, const FConfigureScopeNativeDelegate& OnConfigureScope, ESentryLevel Level = ESentryLevel::Info);
 
 	/**
 	 * Captures a manually created event and sends it to Sentry.
 	 *
 	 * @param Event The event to send to Sentry.
+	 *
+	 * @return Event ID (non-empty if successful)
 	 */
 	UFUNCTION(BlueprintCallable, Category = "Sentry")
-	USentryId* CaptureEvent(USentryEvent* Event);
+	FString CaptureEvent(USentryEvent* Event);
 
 	/**
 	 * Captures a manually created event and sends it to Sentry.
@@ -140,18 +143,16 @@ public:
 	 * @param Event The event to send to Sentry.
 	 * @param OnConfigureScope The callback to configure the scope.
 	 *
-	 * @note: Not supported for Windows/Linux.
+	 * @return Event ID (non-empty if successful)
 	 */
 	UFUNCTION(BlueprintCallable, Category = "Sentry")
-	USentryId* CaptureEventWithScope(USentryEvent* Event, const FConfigureScopeDelegate& OnConfigureScope);
-	USentryId* CaptureEventWithScope(USentryEvent* Event, const FConfigureScopeNativeDelegate& OnConfigureScope);
+	FString CaptureEventWithScope(USentryEvent* Event, const FConfigureScopeDelegate& OnConfigureScope);
+	FString CaptureEventWithScope(USentryEvent* Event, const FConfigureScopeNativeDelegate& OnConfigureScope);
 
 	/**
 	 * Captures a user feedback.
 	 *
 	 * @param UserFeedback The user feedback to send to Sentry.
-	 *
-	 * @note: Not supported for Windows/Linux.
 	 */
 	UFUNCTION(BlueprintCallable, Category = "Sentry")
 	void CaptureUserFeedback(USentryUserFeedback* UserFeedback);
@@ -163,11 +164,9 @@ public:
 	 * @param Email The user email.
 	 * @param Comments The user comments.
 	 * @param Name The optional username.
-	 *
-	 * @note: Not supported for Windows/Linux.
 	 */
 	UFUNCTION(BlueprintCallable, Category = "Sentry")
-	void CaptureUserFeedbackWithParams(USentryId* EventId, const FString& Email, const FString& Comments, const FString& Name);
+	void CaptureUserFeedbackWithParams(const FString& EventId, const FString& Email, const FString& Comments, const FString& Name);
 
 	/**
 	 * Sets a user for the current scope.
@@ -293,7 +292,7 @@ public:
 
 	/** Checks if Sentry event capturing is supported for current settings. */
 	UFUNCTION(BlueprintCallable, Category = "Sentry")
-	bool IsSupportedForCurrentSettings();
+	bool IsSupportedForCurrentSettings() const;
 
 private:
 	/** Adds default context data for all events captured by Sentry SDK. */
@@ -315,25 +314,25 @@ private:
 	void DisableAutomaticBreadcrumbs();
 
 	/** Check whether the event capturing should be enabled for the current build configuration */
-	bool IsCurrentBuildConfigurationEnabled();
+	bool IsCurrentBuildConfigurationEnabled() const;
 
 	/** Check whether the event capturing should be enabled for the current build target */
-	bool IsCurrentBuildTargetEnabled();
+	bool IsCurrentBuildTargetEnabled() const;
 
 	/** Check whether the event capturing should be enabled for promoted builds only */
-	bool IsPromotedBuildsOnlyEnabled();
+	bool IsPromotedBuildsOnlyEnabled() const;
 
 	/** Add custom Sentry output device to intercept logs */
 	void ConfigureOutputDevice();
 
 	/** Add custom Sentry output device to intercept errors */
-	void ConfigureOutputDeviceError();
+	void ConfigureErrorOutputDevice();
 
 private:
 	TSharedPtr<ISentrySubsystem> SubsystemNativeImpl;
 
 	TSharedPtr<FSentryOutputDevice> OutputDevice;
-	TSharedPtr<FSentryOutputDeviceError> OutputDeviceError;
+	TSharedPtr<FSentryErrorOutputDevice> OutputDeviceError;
 
 	UPROPERTY()
 	USentryBeforeSendHandler* BeforeSendHandler;